requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altgraph==0.17.4
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==41.0.2
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.5.0
multidict==6.1.0
packaging==24.2
pefile==2023.2.7
propcache==0.2.0
psutil==6.1.1
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PySide2==5.15.2.1
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
shiboken2==5.15.2.1
tencentcloud-sdk-python==3.0.1296
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
yarl==1.15.2
zipp==3.20.2
